Made group finder component responsive at all device widths
The inline form in groupfinder component was wrapping in an way that
was not aesthetically pleasing.  Played around with sizing so that form
elements now line up appropriately and element widths at different
screen sizes look good.

Also moved the header 'Find groups working towards...' into the
groupfinder element.  I feel that this more closely demonstrates to the
user the utility of the 'Enter your postcode' form.  Previously, with
the header above the groupfinder component, there was a likelihood the
user would fail to associate the form with the header.